+++ src/common/services/record/index.ts
@@ -1,2 +1,0 @@
-export { SchemaTraverser } from './schemaTraverser';
-export { RecordGenerator } from './record';