+++ src/test/__tests__/common/api/profiles.api.test.ts
@@ -1,36 +1,7 @@
 import baseApi from '@common/api/base.api';
-import { fetchProfiles, fetchProfile } from '@common/api/profiles.api';
+import { fetchProfile } from '@common/api/profiles.api';
 
 describe('profiles.api', () => {
-  test('fetchProfiles - calls "baseApi.getJson" and returns profile data', async () => {
-    const testResult = {
-      data: {
-        profiles: [
-          {
-            id: 'profileId_1',
-            name: 'Profile 1',
-            json: {
-              Profile: {
-                resourceTemplates: [
-                  {
-                    id: 'templateId_1',
-                    name: 'Template 1',
-                  },
-                ],
-              },
-            },
-          },
-        ],
-      },
-    };
-    jest.spyOn(baseApi, 'getJson').mockResolvedValue(testResult);
-
-    const result = await fetchProfiles();
-
-    expect(baseApi.getJson).toHaveBeenCalledTimes(1);
-    expect(result).toEqual(testResult);
-  });
-
   test('fetchProfile - calls "baseApi.getJson" with profile ID and returns profile data', async () => {
     const testResult = {
       data: {