+++ src/test/__tests__/common/helpers/bibframe.helper.test.ts
@@ -1,85 +1,53 @@
 import * as BibframeHelper from '@src/common/helpers/bibframe.helper';
-import * as BibframeMappingConstants from '@src/common/constants/bibframeMapping.constants';
-import { getMockedImportedConstant } from '@src/test/__mocks__/common/constants/constants.mock';
+import { QueryParams, ROUTES } from '@src/common/constants/routes.constants';
+import { RecordEditActions } from '@src/common/constants/record.constants';
 
 describe('bibframe.helper', () => {
-  const mockImportedConstant = getMockedImportedConstant(BibframeMappingConstants, 'BF2_TO_BFLITE_MAP');
+  describe('getEditActionPrefix', () => {
+    it('returns RecordEditActions.New when route is RESOURCE_CREATE and no CloneOf query param is present', () => {
+      const route = ROUTES.RESOURCE_CREATE.uri;
+      const search = new URLSearchParams();
 
-  describe('getMappedBFLiteUri', () => {
-    test('returns undefined', () => {
-      mockImportedConstant({
-        testBF20Uri_1: 'testBFLiteUri_1',
-        testBF20Uri_2: 'testBFLiteUri_2',
-      });
-      const uri = 'testBF20Uri_3';
+      const result = BibframeHelper.getEditActionPrefix(route, search);
 
-      const result = BibframeHelper.getMappedBFLiteUri(uri);
-
-      expect(result).toBeUndefined();
+      expect(result).toBe(RecordEditActions.New);
     });
 
-    test('returns mapped URI for string', () => {
-      mockImportedConstant({
-        testBF20Uri_1: 'testBFLiteUri_1',
-      });
-      const uri = 'testBF20Uri_1';
+    it('returns RecordEditActions.Duplicate when route is RESOURCE_CREATE and CloneOf query param is present', () => {
+      const route = ROUTES.RESOURCE_CREATE.uri;
+      const search = new URLSearchParams();
+      search.set(QueryParams.CloneOf, 'some-resource-id');
 
-      const result = BibframeHelper.getMappedBFLiteUri(uri);
+      const result = BibframeHelper.getEditActionPrefix(route, search);
 
-      expect(result).toBe('testBFLiteUri_1');
+      expect(result).toBe(RecordEditActions.Duplicate);
     });
 
-    test('returns mapped URI for object', () => {
-      mockImportedConstant({
-        testBF20Uri_1: {
-          testBF20NestedUri_1: 'testBFLiteNestedUri_1',
-        },
-      });
-      const uri = 'testBF20Uri_1';
-      const path = ['pathItem_1'];
-      const schema = new Map([
-        ['pathItem_1', { uri: 'testBF20NestedUri_1' } as SchemaEntry],
-        ['pathItem_2', { uri: 'testBF20NestedUri_2' } as SchemaEntry],
-      ]);
+    it('returns RecordEditActions.Edit when route is not RESOURCE_CREATE', () => {
+      const route = ROUTES.RESOURCE_EDIT.uri;
+      const search = new URLSearchParams();
 
-      const result = BibframeHelper.getMappedBFLiteUri(uri, schema, path);
+      const result = BibframeHelper.getEditActionPrefix(route, search);
 
-      expect(result).toBe('testBFLiteNestedUri_1');
-    });
-  });
-
-  describe('getUris', () => {
-    function testGetUris(testResult: Record<string, string | undefined>, mapperUri?: string) {
-      jest.spyOn(BibframeHelper, 'getMappedBFLiteUri').mockReturnValue(mapperUri);
-      const uri = 'testBF20Uri_1';
-      const schema = new Map([['pathItem_1', { uri: 'testBF20NestedUri_1' } as SchemaEntry]]);
-      const path = ['pathItem_1'];
-
-      const result = BibframeHelper.getUris({ uri, schema, path });
-
-      expect(result).toEqual(testResult);
-    }
-
-    test('returns an object with uris', () => {
-      testGetUris({ uriBFLite: 'testBFLiteUri', uriWithSelector: 'testBFLiteUri' }, 'testBFLiteUri');
+      expect(result).toBe(RecordEditActions.Edit);
     });
 
-    test('returns an object with empty uriBFLite', () => {
-      testGetUris({ uriBFLite: undefined, uriWithSelector: 'testBF20Uri_1' });
+    it('returns RecordEditActions.Edit when route is undefined', () => {
+      const route = undefined;
+      const search = new URLSearchParams();
+
+      const result = BibframeHelper.getEditActionPrefix(route, search);
+
+      expect(result).toBe(RecordEditActions.Edit);
     });
 
-    test('returns an object for duplicate bf2.0 uris', () => {
-      const duplicateUri = 'http://id.loc.gov/ontologies/bibframe/contribution';
-      const nextUri = 'http://id.loc.gov/ontologies/bibframe/Contribution';
-      const resultUri = 'http://bibfra.me/vocab/lite/contributor';
+    it('returns RecordEditActions.New when search params are undefined', () => {
+      const route = ROUTES.RESOURCE_CREATE.uri;
+      const search = undefined;
 
-      const result = BibframeHelper.getUris({ uri: duplicateUri, dataTypeURI: nextUri, schema: new Map() });
-      const testResult = {
-        uriBFLite: resultUri,
-        uriWithSelector: resultUri,
-      };
+      const result = BibframeHelper.getEditActionPrefix(route, search);
 
-      expect(result).toEqual(testResult);
+      expect(result).toBe(RecordEditActions.New);
     });
   });
 });